requirements.txt
PySide6>=6.5.0
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
PyYAML>=6.0.0